requirements.txt
-e .
accelerate==1.6.0
bandit==1.8.3
black==25.1.0
faiss-cpu==1.11.0
flake8==7.2.0
gradio==5.28.0
hf_xet==1.1.0
langchain==0.3.24
langchain-community>=0.3.23
langchain-openai==0.3.15
# boto3==2.49.0
langchain_huggingface==0.1.2
# ipykernel==6.29.5
loguru==0.7.3
mypy==1.15.0
nltk==3.9.1
openai==1.76.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
pillow==11.2.1
plotly==6.0.1
pre-commit==4.2.0
# pytorch is installed using cuda
pydantic==2.11.4
PyMuPDF==1.25.5
# tiktoken==0.9.0
python-dotenv==1.1.0
rank_bm25==0.2.2
# pydub==0.25.1
requests==2.32.3
seaborn==0.13.2
streamlit==1.45.0
toml==0.10.2
transformers==4.51.3
umap-learn==0.5.7
